update my selected crops test
to use api data, move tests to filter file
