+++ cypress/e2e/filters-neccc.cy.js
@@ -1,7 +1,6 @@
 /* eslint-disable */
-// import { necccFilterTests } from '../support/e2e-neccc';
-import { flipCoverCropName } from "../../src/shared/constants";
-import { testFiltersByType } from "../support/e2e";
+import { flipCoverCropName } from '../../src/shared/constants';
+import { testFiltersByType } from '../support/e2e';
 
 /* eslint-disable no-undef */
 
@@ -40,22 +39,22 @@ describe('Test all possible interactions on the NECCC Crop Calendar Page', () =>
       });
       cy.wait('@getFilters').then((interception) => {
         const { data } = interception.response.body;
-        Cypress.env('filters', data);
+        // 'Goals' are for goals not filters, remove the data from it
+        const filterTypes = data.map((filter) => filter.label).filter((label) => label !== 'Goals');
+        const allFilters = data.reduce((res, filter) => {
+          if (filter.label === 'Goals') return res;
+          return [...res, ...filter.attributes];
+        }, []);
+        Cypress.env('filterTypes', filterTypes);
+        Cypress.env('allFilters', allFilters);
       });
     });
   });
 
-  // necccFilterTests();
-  it.only('should work on all types of filters', () => {
+  it('should work on all types of filters', () => {
     const cropData = Cypress.env('cropData');
-    const filters = Cypress.env('filters');
-    // 'Goals' are for goals not filters, remove the data from it
-    const filterTypes = filters.map(filter => filter.label).filter((label) => label !== 'Goals');
-    
-    const allFilters = filters.reduce((res, filter) => {
-      if (filter.label === 'Goals') return res;
-      return [...res, ...filter.attributes];
-    }, [])
+    const filterTypes = Cypress.env('filterTypes');
+    const allFilters = Cypress.env('allFilters');
 
     const allFilterDataTypes = allFilters.reduce((res, filter) => {
       const dataType = filter.dataType.label;
@@ -73,7 +72,7 @@ describe('Test all possible interactions on the NECCC Crop Calendar Page', () =>
     const testFilterValues = allFilterDataTypes.reduce((res, dataType) => {
       const filter = allFilters.find((filter) => filter.dataType.label === dataType);
       if (!filter) return res;
-      return [...res, filter.values.map((v) => v.dataType === "number" && filter.dataType.label !== "currency" ? parseInt(v.value) : v.value)];
+      return [...res, filter.values.map((v) => (v.dataType === 'number' && filter.dataType.label !== 'currency' ? parseInt(v.value) : v.value))];
     }, []);
 
     const filterResults = {};
@@ -81,18 +80,18 @@ describe('Test all possible interactions on the NECCC Crop Calendar Page', () =>
       const result = {};
       testFilterValues[i].forEach((value) => {
         result[value] = [];
-      })
+      });
       cropData.forEach((crop) => {
         const attr = crop.attributes.find((attr) => attr.label === filter);
         if (attr) {
           attr.values.forEach((val) => {
             const { value } = val;
             result[value] = [...result[value], flipCoverCropName(crop.label)];
-          })
+          });
         }
-      })
+      });
       filterResults[filter] = result;
-    })
+    });
     cy.log(filterTypes, allFilterDataTypes, testFilters, testFilterValues, filterResults);
     testFiltersByType(filterTypes, testFilters, testFilterValues, filterResults);
   });
@@ -116,5 +115,126 @@ describe('Test all possible interactions on the NECCC Crop Calendar Page', () =>
             });
           });
       });
+  });
+
+  it('should display selected crops number on My Select Crops tab and show corresponding crop cards in My Selected Crops', () => {
+    const btnIdx = [0, 1, 2];
+    const cropLabels = [];
+    const cardLabels = [];
+
+    btnIdx.forEach((idx) => {
+      cy.getByTestId(`cart-btn-${idx}`).click({ force: true });
+      cy.get(`[data-test=crop-list-tr-${idx}]`).then(($row) => {
+        cy.wrap($row)
+          .find('[data-test=crop-calendar-crop-name]')
+          .should('be.visible')
+          .then(($label) => {
+            cropLabels.push($label.text());
+          });
+      });
+    });
+
+    btnIdx.forEach((idx) => {
+      cy.get(`[data-test=delete-forever-icon-${idx}]`).should('exist');
+      cy.get(`[data-test=add-circle-outline-icon-${idx}]`).should('not.exist');
+    });
+
+    cy.get('[data-test=badge] .MuiBadge-badge') // Select all badges
+      .each(($badge) => {
+        cy.wrap($badge).should('have.text', btnIdx.length); // Assert that each badge has the text '2'
+      });
+
+    cy.log('===CROP LABELS===', cropLabels);
+
+    cy.get("[data-test='my selected crops-btn']")
+      .first()
+      .click()
+      .then(() => {
+        btnIdx.forEach((idx) => {
+          cy.assertByTestId(`crop-card-${idx}`);
+        });
+        cy.get('[data-test^=crop-card-label-]').each(($label) => {
+          // Ensure the label is visible
+          cy.wrap($label).should('be.visible')
+            .invoke('text')
+            .then((labelText) => {
+              cardLabels.push(flipCoverCropName(labelText));
+              cy.log(`Crop Card Label: ${labelText}`);
+            });
+        })
+          .then(() => {
+            cy.log('=== CARD LABELS===', cardLabels);
+            expect(cropLabels.length).to.equal(cardLabels.length);
+            btnIdx.forEach((idx) => {
+              expect(cropLabels[idx].trim().toLowerCase()).to.equal(cardLabels[idx].trim().toLowerCase());
+            });
+          });
+      });
+  });
+
+  describe('tests for My Selected Crops filters', () => {
+    beforeEach(() => {
+      // select top 3 crops and go to My Select Crops page, open all accordions
+      const btnIdx = [0, 1, 2];
+      btnIdx.forEach((idx) => {
+        cy.getByTestId(`cart-btn-${idx}`).click();
+      });
+      cy.get("[data-test='my selected crops-btn']")
+        .first()
+        .click({ force: true })
+        .then(() => {
+          cy.get("[data-test='comparison-view-btn']")
+            .should('be.visible')
+            .click();
+        });
+
+      const filterTypes = Cypress.env('filterTypes');
+      for (let i = 0; i < filterTypes.length; i++) {
+        // FIXME: need verification for this
+        if (filterTypes[i] === 'Soil Conditions') continue;
+        cy.assertByTestId(`"${filterTypes[i].toUpperCase()}"`);
+        cy.getByTestId(`${filterTypes[i].toUpperCase()}-expandmore-icon`).click();
+      }
+    });
+
+    it('should show all filters and comparisons', () => {
+      const allFilters = Cypress.env('allFilters');
+      const allFilterName = allFilters.map((filter) => filter.label);
+
+      for (let i = 0; i < allFilterName.length; i++) {
+        // FIXME: need verification for this
+        if (allFilterName[i] === 'Supports Mycorrhizae') continue;
+
+        cy.get(`[data-test='${allFilterName[i]}-checkbox']`)
+          .click()
+          .then(() => {
+            cy.assertByTestId(`"${allFilterName[i]}-row"`);
+          });
+
+        cy.get(`[data-test='${allFilterName[i]}-checkbox']`)
+          .click()
+          .then(() => {
+            cy.get(`[data-test="${allFilterName[i]}-row"]`).should('not.exist');
+          });
+      }
+    });
+
+    it('should be able to use Show All and Clear Variables button to control comparisons', () => {
+      const allFilters = Cypress.env('allFilters');
+      const allFilterName = allFilters.map((filter) => filter.label);
+
+      cy.getByTestId('my-selected-crops-show-all').click();
+      allFilterName.forEach((filter) => {
+        // FIXME: need verification for this
+        if (filter === 'Supports Mycorrhizae') return;
+        cy.assertByTestId(`"${filter}-row"`);
+      });
+      cy.getByTestId('my-selected-crops-clear-variables').click();
+      allFilterName.forEach((filter) => {
+        // FIXME: need verification for this
+        if (filter === 'Supports Mycorrhizae') return;
+        cy.getByTestId(`${filter}-rows`).should('not.exist');
+      });
+    });
   });
 });
