+++ cypress/support/e2e.js
@@ -1,6 +1,5 @@
 /* eslint-disable */
 /* eslint-disable no-undef */
-import { flipCoverCropName } from '../../src/shared/constants';
 import './commands';
 
 const rowOpacity = 0.55;
@@ -28,7 +27,6 @@ export const testFiltersByType = (filterTypes, testFilters, testFilterValues, fi
     cy.getByTestId('crop-side-bar-clear-filters').click();
   });
 };
-
 
 Cypress.Commands.add('testFilters', ({
   // eslint-disable-next-line no-unused-vars
@@ -89,100 +87,5 @@ const checkRows = (filterType, filterIndex, filterResult) => {
           });
         });
       });
-  });
-};
-
-export const checkComparisonTableRows = ({ filterName }) => {
-  cy.get(`[data-test='${filterName}-checkbox']`)
-    .click()
-    .then(() => {
-      cy.assertByTestId(`"${filterName}-row"`);
-    });
-
-  cy.get(`[data-test='${filterName}-checkbox']`)
-    .click()
-    .then(() => {
-      cy.get(`[data-test="${filterName}-row"]`).should('not.exist');
-    });
-};
-
-export const mySelectedCropsCommonTests = () => {
-  it('should display selected crops number on My Select Crops tab and show corresponding crop cards in My Selected Crops', () => {
-    const btnIdx = [0, 1, 2];
-    const cropLabels = [];
-    const cardLabels = [];
-
-    btnIdx.forEach((idx) => {
-      cy.assertByTestId(`cart-btn-${idx}`).click({ force: true });
-      cy.get(`[data-test=crop-list-tr-${idx}]`).then(($row) => {
-        cy.wrap($row)
-          .find('[data-test=crop-calendar-crop-name]')
-          .should('be.visible')
-          .then(($label) => {
-            cropLabels.push($label.text());
-          });
-      });
-    });
-
-    btnIdx.forEach((idx) => {
-      cy.get(`[data-test=delete-forever-icon-${idx}]`).should('exist');
-      cy.get(`[data-test=add-circle-outline-icon-${idx}]`).should('not.exist');
-    });
-
-    cy.get('[data-test=badge] .MuiBadge-badge') // Select all badges
-      .each(($badge) => {
-        cy.wrap($badge).should('have.text', btnIdx.length); // Assert that each badge has the text '2'
-      });
-
-    cy.log('===CROP LABELS===', cropLabels);
-
-    cy.get("[data-test='my selected crops-btn']")
-      .first()
-      .click()
-      .then(() => {
-        btnIdx.forEach((idx) => {
-          cy.assertByTestId(`crop-card-${idx}`);
-        });
-        cy.get('[data-test^=crop-card-label-]').each(($label) => {
-          // Ensure the label is visible
-          cy.wrap($label).should('be.visible')
-            .invoke('text')
-            .then((labelText) => {
-              cardLabels.push(flipCoverCropName(labelText));
-              cy.log(`Crop Card Label: ${labelText}`);
-            });
-        })
-          .then(() => {
-            cy.log('=== CARD LABELS===', cardLabels);
-            expect(cropLabels.length).to.equal(cardLabels.length);
-            btnIdx.forEach((idx) => {
-              expect(cropLabels[idx].trim().toLowerCase()).to.equal(cardLabels[idx].trim().toLowerCase());
-            });
-          });
-      });
-  });
-};
-
-export const presenceOfFiltersTests = (sidebarFilters) => {
-  it('should have presence of all the filters', () => {
-    const btnIdx = [0, 1, 2];
-    btnIdx.forEach((idx) => {
-      cy.assertByTestId(`cart-btn-${idx}`).click({ force: true });
-    });
-    cy.get("[data-test='my selected crops-btn']")
-      .first()
-      .click({ force: true })
-      .then(() => {
-        cy.get("[data-test='comparison-view-btn']")
-          .should('be.visible')
-          .click();
-      });
-
-    for (let i = 0; i < sidebarFilters.length; i++) {
-      // eslint-disable-next-line no-continue
-      if (sidebarFilters[i] === 'SOIL CONDITIONS') continue;
-      cy.assertByTestId(`"${sidebarFilters[i]}"`);
-      cy.assertByTestId(`"${sidebarFilters[i]}-expandmore-icon"`);
-    }
   });
 };
